pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = ""
requires-python = ">=3.11"

dependencies = [
  "Django>=5.0,<5.2",
  "psycopg2-binary>=2.9.9",
  "django-bootstrap5>=23.4",
  "django-filter>=23.5",
  "python-dotenv>=1.0",
  "rollbar>=1.0",
  "gunicorn>=22.0",
  "dj-database-url>=2.1 ; sys_platform != 'emscripten'",
  "beautifulsoup4>=4.14.2",
  "requests>=2.32.5",
]

[dependency-groups]
dev = [
  "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["task_manager"]
